+++ app/services/import_actuals.rb
@@ -1,3 +1,4 @@
+# This originally handled only Actuals, but now also handles Refunds as well.
 class ImportActuals
   Error = Struct.new(:row, :column, :value, :message) {
     def csv_row
@@ -38,8 +39,8 @@ class ImportActuals
     importer.actual
   end
 
-  def add_error(row_number, column, value, message)
-    @errors << Error.new(row_number, Converter::FIELDS[column], value, message)
+  def add_error(row_number, column, cell_value, message)
+    @errors << Error.new(row_number, Converter::FIELDS[column], cell_value, message)
   end
 
   class RowImporter
@@ -54,7 +55,7 @@ class ImportActuals
 
     def import_row
       @converter = Converter.new(@row)
-      return if @converter.zero_value_actual?
+      return if @converter.zero_value?
 
       @errors.update(@converter.errors)
 
@@ -88,8 +89,12 @@ class ImportActuals
       attrs = @converter.to_h
       assign_default_values(attrs)
 
-      creator = CreateActual.new(activity: @activity, report: @report)
-      result = creator.call(attributes: attrs)
+      creator = if @converter.actual_class == Refund
+        CreateRefund.new(activity: @activity, report: @report, user: @uploader)
+      else
+        CreateActual.new(activity: @activity, report: @report, user: @uploader)
+      end
+      result = creator.call(attributes: attrs.except(:refund_value, :actual_value))
       return unless result
 
       result.object.errors.each do |error|
@@ -113,12 +118,15 @@ class ImportActuals
       activity: "Activity RODA Identifier",
       financial_quarter: "Financial Quarter",
       financial_year: "Financial Year",
-      value: "Value",
+      actual_value: "Actual Value",
+      refund_value: "Refund Value",
       receiving_organisation_name: "Receiving Organisation Name",
       receiving_organisation_type: "Receiving Organisation Type",
       receiving_organisation_reference: "Receiving Organisation IATI Reference",
       comment: "Comment"
     }
+
+    NON_VALUE_FIELDS = FIELDS.without :actual_value, :refund_value
 
     attr_reader :activity, :errors
 
@@ -127,6 +135,14 @@ class ImportActuals
       @errors = {}
       @attributes = convert_to_attributes
       @activity = @attributes.delete(:activity)
+    end
+
+    # We must have precisely one value, and we will check if
+    # the other is empty when we validate the value it contains
+    def actual_class
+      return Actual if @row["Refund Value"].nil?
+      return Refund if @row["Actual Value"].nil? || @row["Actual Value"] == "0.00"
+      @errors[:actual_value] = [@row["Actual Value"], I18n.t("importer.errors.actual.non_numeric_value")]
     end
 
     def raw(attr_name)
@@ -138,19 +154,28 @@ class ImportActuals
     end
 
     def convert_to_attributes
-      FIELDS.each_with_object({}) do |(attr_name, column_name), attrs|
+      attrs = {value: convert_values}
+      NON_VALUE_FIELDS.each_with_object(attrs) do |(attr_name, column_name), attrs|
         attrs[attr_name] = convert_to_attribute(attr_name, @row[column_name])
       end
     end
 
-    def convert_to_attribute(attr_name, value)
-      original_value = value.clone
-      value = value.to_s.strip
+    def convert_values
+      if actual_class == Actual
+        convert_to_attribute(:actual_value, @row["Actual Value"])
+      else
+        convert_to_attribute(:refund_value, @row["Refund Value"])
+      end
+    end
+
+    def convert_to_attribute(attr_name, attr_value)
+      original_value = attr_value.clone
+      attr_value = attr_value.to_s.strip
 
       converter = "convert_#{attr_name}"
-      value = __send__(converter, value) if respond_to?(converter)
+      attr_value = __send__(converter, attr_value) if respond_to?(converter)
 
-      value
+      attr_value
     rescue Encoding::CompatibilityError
       @errors[attr_name] = [
         original_value.force_encoding("UTF-8"),
@@ -162,19 +187,18 @@ class ImportActuals
       nil
     end
 
-    def convert_comment(body)
-      Comment.new(body: body) if body.present?
-    end
-
     def convert_activity(id)
       Activity.by_roda_identifier(id)
     end
 
-    def convert_value(value)
+    def convert_financial_value(value)
       ConvertFinancialValue.new.convert(value)
     rescue ConvertFinancialValue::Error
       raise I18n.t("importer.errors.actual.non_numeric_value")
     end
+
+    alias_method :convert_actual_value, :convert_financial_value
+    alias_method :convert_refund_value, :convert_financial_value
 
     def convert_receiving_organisation_type(type)
       validate_from_codelist(
@@ -200,7 +224,7 @@ class ImportActuals
       code
     end
 
-    def zero_value_actual?
+    def zero_value?
       @attributes[:value].present? && @attributes[:value].zero?
     end
   end